Install foreman, adding example .env and Procfile files.
foreman is a utility which enables applications with multiple
components (such as the web-app and resque workers) to be booted
in development, complete with environment variables set.

tunefl uses environment variables instead of additional config files,
so some such approach is most useful.

An example .env file is provided as .env.example; this can be
copied to .env, and adjusted as appropriate. Similarly for the
example Procfile.example, to be copied to Procfile. The settings
present in the example are such that no further configuration is
required after copying, if file-storage is going to be used without
a CDN.

To start tunefl in development, booting both the web-app and a worker:

    foreman start