[#2] Fix Score.blocked accessible.
Score#new was called with the params array. However, Score.blocked is
marked as accessible, for ActiveAdmin. Changed to explicitly set each
parameter required from the params array - which is arguably a safer
approach.

Use rescue nil method, which catches an error on that line - extremely
likely to be a non-existent params path. The attribute then doesn't get
reassigned, because of ||= .